# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "API documentation for Laravel project",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a list of all reservations",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "List of all reservations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "reservation_date",
                                    "return_date"
                                ],
                                "properties": {
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-01T12:00:00Z"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-10T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get details of a specific reservation",
                "operationId": "a226b43b2a6a846245e658d01bf00720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update an existing reservation",
                "operationId": "02fa683a9f80a0563596fb06685bcbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "car_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-05T14:00:00Z"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-15T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "title": "Car",
                "description": "A car available for reservation",
                "properties": {
                    "id": {
                        "description": "Car ID",
                        "type": "integer",
                        "example": 1
                    },
                    "make": {
                        "description": "Car brand",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "description": "Car model",
                        "type": "string",
                        "example": "Corolla"
                    },
                    "year": {
                        "description": "Manufacturing year",
                        "type": "integer",
                        "example": 2023
                    },
                    "is_available": {
                        "description": "Availability status",
                        "type": "boolean",
                        "example": true
                    },
                    "pricing": {
                        "$ref": "#/components/schemas/CarPricing"
                    },
                    "specs": {
                        "$ref": "#/components/schemas/CarSpec"
                    },
                    "photos": {
                        "description": "List of car photos",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CarPhoto"
                        }
                    },
                    "reservations": {
                        "description": "List of reservations for this car",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Reservation"
                        }
                    }
                },
                "type": "object"
            },
            "CarPhoto": {
                "title": "Car Photo",
                "description": "Photo associated with a car",
                "properties": {
                    "id": {
                        "description": "Photo ID",
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "description": "Car ID",
                        "type": "integer",
                        "example": 1
                    },
                    "photo_name": {
                        "description": "Car photo filename",
                        "type": "string",
                        "example": "toyota-corolla.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "CarPricing": {
                "title": "Car Pricing",
                "description": "Pricing details of a car",
                "properties": {
                    "id": {
                        "description": "Pricing ID",
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "description": "Car ID",
                        "type": "integer",
                        "example": 1
                    },
                    "hour_price": {
                        "description": "Hourly rental price",
                        "type": "number",
                        "format": "float",
                        "example": 15.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "CarSpec": {
                "title": "Car Specifications",
                "description": "Technical specifications of a car",
                "properties": {
                    "id": {
                        "description": "Car Specification ID",
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "description": "Car ID",
                        "type": "integer",
                        "example": 1
                    },
                    "power": {
                        "description": "Car engine power in HP",
                        "type": "integer",
                        "example": 150
                    },
                    "color": {
                        "description": "Car color",
                        "type": "string",
                        "example": "Red"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-02T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "title": "Reservation",
                "description": "A car reservation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "car_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "reservation_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-01T12:00:00Z"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-10T12:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-02T12:00:00Z"
                    },
                    "car": {
                        "$ref": "#/components/schemas/Car"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}